Migrate TagRegistration to TagTemplate
