+++ src/main/java/org/spongepowered/api/datapack/DataPackType.java
@@ -28,8 +28,7 @@ import io.leangen.geantyref.TypeToken;
 import org.spongepowered.api.advancement.Advancement;
 import org.spongepowered.api.event.lifecycle.RegisterDataPackValueEvent;
 import org.spongepowered.api.item.recipe.RecipeRegistration;
-import org.spongepowered.api.tag.Tag;
-import org.spongepowered.api.tag.TagRegistration;
+import org.spongepowered.api.tag.TagTemplate;
 import org.spongepowered.api.util.annotation.CatalogedBy;
 import org.spongepowered.api.world.WorldTypeTemplate;
 import org.spongepowered.api.world.server.WorldTemplate;
@@ -60,6 +59,6 @@ public interface DataPackType<T> {
 
         DataPackType<WorldTemplate> world();
 
-        DataPackType<TagRegistration> tag();
+        DataPackType<TagTemplate> tag();
     }
 }
