+++ src/main/java/org/spongepowered/api/tag/Tag.java
@@ -24,11 +24,7 @@
  */
 package org.spongepowered.api.tag;
 
-import org.checkerframework.checker.nullness.qual.NonNull;
-import org.spongepowered.api.ResourceKey;
 import org.spongepowered.api.ResourceKeyed;
-import org.spongepowered.api.Sponge;
-import org.spongepowered.api.data.Key;
 import org.spongepowered.api.registry.DefaultedRegistryValue;
 
 import java.util.Collection;
@@ -39,17 +35,7 @@ import java.util.Collection;
  *
  * @param <T> Type that this tag groups
  */
-public interface Tag<T> extends DefaultedRegistryValue, ResourceKeyed, TagRegistration {
-
-    /**
-     * Creates a builder to create a new {@link Tag}
-     * or modify an existing one.
-     * @return Builder to build a new Tag.
-     */
-    @SuppressWarnings("unchecked")
-    static Builder<?> builder() {
-        return Sponge.game().builderProvider().provide(Builder.class);
-    }
+public interface Tag<T> extends DefaultedRegistryValue, ResourceKeyed, TagTemplate {
 
     /**
      * Get every value in this Tag.
@@ -65,61 +51,4 @@ public interface Tag<T> extends DefaultedRegistryValue, ResourceKeyed, TagRegist
      * @return Whether the value is contained in this tag.
      */
     boolean contains(T value);
-
-    interface Builder<T extends Taggable> extends org.spongepowered.api.util.ResourceKeyedBuilder<TagRegistration, Builder<T>> {
-
-        /**
-         * Sets the tag type and generifies the builder.
-         *
-         * @param tagType Type to build a tag for, e.g {@link TagTypes#BLOCK_TYPE}
-         * @return This builder, for chaining
-         */
-        <NT extends Taggable> Builder<NT> type(TagType<NT> tagType);
-
-        /**
-         * <p>Whether to replace instead of append if the tag already
-         * exists.</p>
-         *
-         * <p>By default, if the {@link ResourceKey#value()} is the same as
-         * another data pack / plugin, the two will be combined, and their
-         * tags appended to each other.</p>
-         *
-         * @param replace Whether to replace instead of merge.
-         * @return This builder, for chaining.
-         */
-        Builder<T> replace(boolean replace);
-
-        default Builder<T> replace() {
-            return replace(true);
-        }
-
-        /**
-         * Adds a child tag to this tag.
-         * This adds all of its values to this tag.
-         *
-         * @param childTag Child tag to add.
-         * @return This builder, for chaining
-         */
-        Builder<T> addChild(Tag<T> childTag);
-
-        Builder<T> addChildren(Collection<Tag<T>> children);
-
-        /**
-         * Adds the value to the tag.
-         *
-         * @param value Value to add
-         * @return This builder, for chaining
-         */
-        Builder<T> addValue(T value);
-
-        Builder<T> addValues(Collection<T> values);
-
-        /**
-         * A nicely generified tag, <b>That must be registered</b> in {@link org.spongepowered.api.event.lifecycle.RegisterDataPackValueEvent}.
-         *
-         * @return The built tag.
-         */
-        @Override
-        @NonNull TagRegistration build();
-    }
 }
